Updated logging system and added daytime check feature
Added a new feature to check if it is daytime and adjust logging accordingly. Also enhanced logging system by integrating Serilog for better readability and control. New settings were added in AppSettings for controlling the logging interval and enabling/disabling the daytime check. Also, added necessary configuration for Serilog in appsettings.json. Package references relevant to these changes were updated in SofarKTLXLogger.csproj. Introduced a Daytime service and implemented it in Logger.cs and SolarmanV5Client.cs. The logic for these changes is distributed to respective methods to maintain coherence and readability.