Add detailed logging and refine data handling in SolarLogger application
Introduced a docker-compose-example for deploying the application, while refining handling of InverterData and Logger classes to properly process and log data from solar inverters. This includes splitting and catching socket exception separately for InverterData and PvData in Logger.cs, along with overhauling log points for writing data to InfluxDB.

Incoming and outgoing byte data from SolarmanV5Client are also being logged for better debugging. Adjustments are made in DaytimeService.cs to log critical times and correctly account for timezones during sunrise and sunset.

Additional logging levels are displayed in Program.cs for reference. Updates to .gitignore and .dockerignore are made to ignore new 'sample-data' and 'Logs' directories.

The changes are necessary to maximize the application's data processing efficiency and ease debugging with detailed logging.