Add CI workflow for auto versioning and deployment
This commit adds a new continuous integration (CI) workflow that allows for automatic versioning and deployment. The workflow includes options for different version levels (major, minor, patch) and a specific version input, as well as an option to push the Docker image to DockerHub. It also includes steps for getting and setting variables such as previous version tag, next version (with semver), commit branch and commit hash. A logger job builds and pushes the Docker image. The workflow ends by creating a new GitHub release unless the version level is 'commit'.