Added Dockerfile and CI workflow for logger deployment
This commit introduces a Dockerfile for building and executing the logger application in a containerized environment. It also includes a sophisticated GitHub Actions workflow `ci2.yml` for automating the build, push and tagging of Docker images with semantic versioning.

A detailed `ci2.yml` file has been added to handle the workflow which includes jobs for building variables, building and pushing the logger, and creating a new release. It uses actions for semantic versioning, Docker setup, QEMU setup, DockerHub login, Docker build-push, and creating a GitHub release.

The Dockerfile ensures that our application runs seamlessly in any environment where Docker is installed, thereby simplifying the process of deployment.

In summary, these changes establish a foundation for continuous integration and deployment (CI/CD) of our logger application.