Add timeout handling and optimize logging in real-time data fetch
In 'Logger.cs', introduced a CancellationToken that gets cancelled after a timeout period for handling long-running operations and to prevent it from hanging indefinitely. This CancellationToken replaces the previous one used in LogRealTimeInformation.

In 'SolarmanV5Client.cs', the LogTrace statement marking the sent bytes was moved after the 'WriteAsync' call. This ensures that the logging message only appears when the data has been successfully sent and not before. This provides more accurate logging.